fastapi_gateway/requirements.txt
# 의존성 
fastapi
uvicorn
redis
pyjwt
requests
cryptography